requirements.txt
aliyun-python-sdk-core-v3==2.8.8
aliyun-python-sdk-dysmsapi==1.0.0
amqp==2.3.2
billiard==3.5.0.3
blinker==1.4
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Mail==0.9.1
flask-paginate==0.5.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
Pillow==5.1.0
pycryptodome==3.6.1
PyMySQL==0.8.1
python-memcached==1.59
pytz==2018.4
qiniu==7.2.2
redis==2.10.6
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.22

vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2
